일부만 검색되는 문제

+++ space/src/main/java/xyz/itwill/dao/MemberHostBoardDAO.java
@@ -61,7 +61,7 @@ public interface MemberHostBoardDAO {
 	List<SelectMember> selectHostQuestionList(Map<String, Object> map);
 	
 	// Filter 
-	int selectStarAverage(int sNo);
+	double selectStarAverage(int sNo);
 	List<Space> selectSpaceSno(Map<String, Object> params);
 	int selectSpaceCount(Map<String, Object> params);
 	List<Space> selectSpaceList(Map<String, Object> map);
